# api-ts/openapi-agent.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Suwappu Agent API",
    "description": "Cross-chain DEX API for AI agents. Swap tokens across 7+ chains via structured endpoints.",
    "version": "0.4.0",
    "contact": {
      "name": "Suwappu",
      "url": "https://suwappu.bot"
    }
  },
  "servers": [
    {
      "url": "https://api.suwappu.bot/v1/agent",
      "description": "Production"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/quote": {
      "post": {
        "summary": "Get a swap quote",
        "operationId": "getQuote",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["from_token", "to_token", "amount"],
                "properties": {
                  "from_token": { "type": "string", "description": "Source token symbol (e.g. ETH, SOL, USDC)" },
                  "to_token": { "type": "string", "description": "Destination token symbol" },
                  "amount": { "type": "string", "description": "Amount in human-readable format (e.g. \"0.5\")" },
                  "chain": { "type": "string", "description": "Chain key (e.g. base, solana)" },
                  "from_chain": { "type": "string" },
                  "to_chain": { "type": "string" },
                  "wallet_address": { "type": "string" },
                  "slippage": { "type": "number", "minimum": 0, "maximum": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Swap quote with quote_id for execution",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "quote_id": { "type": "string" },
                    "from_token": { "type": "object" },
                    "to_token": { "type": "object" },
                    "amount_in": { "type": "string" },
                    "amount_out": { "type": "string" },
                    "exchange_rate": { "type": "string" },
                    "expires_in_seconds": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/swap": {
      "post": {
        "summary": "Get unsigned swap transaction",
        "operationId": "getSwapTransaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet_address"],
                "properties": {
                  "quote_id": { "type": "string" },
                  "wallet_address": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Unsigned transaction for signing" }
        }
      }
    },
    "/swap/execute": {
      "post": {
        "summary": "Execute managed swap via server-side signing",
        "operationId": "executeSwap",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quote_id"],
                "properties": {
                  "quote_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Swap submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "swap_id": { "type": "integer" },
                    "status": { "type": "string" },
                    "tx_hash": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/swap/status/{swapId}": {
      "get": {
        "summary": "Get swap status",
        "operationId": "getSwapStatus",
        "parameters": [
          {
            "name": "swapId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Swap status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "swap_id": { "type": "integer" },
                    "status": { "type": "string" },
                    "tx_hash": { "type": "string", "nullable": true },
                    "from_token": { "type": "string" },
                    "to_token": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/execute": {
      "post": {
        "summary": "Execute natural language command",
        "operationId": "executeCommand",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["command"],
                "properties": {
                  "command": { "type": "string", "maxLength": 500, "description": "Natural language command (e.g. \"swap 0.5 ETH to USDC on Base\")" },
                  "wallet_address": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Command result" }
        }
      }
    },
...
    "/tokens": {
      "get": {
        "summary": "List available tokens",
        "operationId": "listTokens",
        "parameters": [
          { "name": "chain", "in": "query", "schema": { "type": "string" }, "description": "Filter by chain (e.g. base, solana). Omit for all chains." },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Filter by symbol substring (e.g. USD)" }
        ],
        "responses": {
          "200": {
            "description": "Token list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "chain": { "type": "string" },
                    "chain_id": {},
                    "tokens": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "symbol": { "type": "string" },
                          "address": { "type": "string" },
                          "decimals": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from POST /register (suwappu_sk_...)"
      }
    },
...
  }
}